openbsdpkg: multi-pkg for remove/purge, optimizations
This commit brings support for the "pkgs" argument to the remove and
purge functions. In addition, the use of __context__ has been added to
list_pkgs. All functions which modify the package database clear the
__context__ variable before running list_pkgs again to ensure that
changes are detected.

The process by which packages are listed has been simplified, as well.

Missing comparison functions were also added.