Clarify and make connection settings more flexible for libvirt
Previously you could specify a hostname or IP in the minion configuration. It was not flexible enough. Now you can either specify an explicit libvirt URI, or a libvirt connection alias.

Note: this is an incompatible change to #6672 when it comes to the configuration. The connection string can no longer contain a hostname.

 - http://libvirt.org/drvesx.html#uriformat
 - http://libvirt.org/uri.html#URI_config