Fix yaml.load refs
These are not really security issues because Salt parses all
incomming arguments via yaml.safe_load before they get here, these
really are not needed. I kept the yaml check in for pillar.ext
simply because we want to be able to support an internal application
firing the routine with a yaml sting.