Add registry argument to call to npm.install
In #13033, I removed this argument from the call to npm.list, it appears
it was mistakenly put there instead of npm.install. Removing the
argument triggered a lint failure:

https://github.com/saltstack/salt/issues/13065

This commit adds the argument to the call to npm.install, which should
allow the state to work as originally intended.