Consolidating the _get_options function from all the returners that were using it into the config module.  Additionally some cleanup and ability to pass defaults to the get_returner_options function.
fixing lint issues

moving the get_options function into a utils.py under returners.  previous attempt at using a call to the salt config module doesn't work if the returner is called from the scheduler.

adding missing docstring

Unit tests were failing for SMTP returner because __opts__ was not available.  Ensuring it's available, empty, if it doesn't already exist.

removing the check and population with an empty dict for __salt__ and __opts__.  Adding an empty __opts__ to the smtp returner unit tests.