Fix issue 22306
Be aware of when we download files from S3 that have an environment
prefix. For example, if the S3 bucket contains /base/top.sls then the
file will be downloaded as {pillar_dir}/base/top.sls rather than
{pillar_dir}/top.sls which will mean that we are building our pillar
with the wrong root and won't actually ingest any pillar data.

Now we take into account whether or not we are in multiple_env mode and
correct the pillar root accordingly.

We also need to avoid recursively recreating the pillar when we compile
the pillar. There may be a more canonical way to do this, I'm not
familiar enough with salt internals to know, but by ripping out the s3
config from the ext_pillar config we avoid the infinite recursion.

May be related to 18355