s3fs breaks when fetching files from s3
Per https://github.com/saltstack/salt/issues/23567 there is an exception
when trying to parse headers.  The values in the returned ret['headers']
now defaults to just being a list of strings containing the keys, but no
values are included.  It seems that in the past a list of strings of
key:value pairs was returned.

This patch requests full_headers, which instead returns a dictionary of
header keys and values.  The keys are lowercased, but otherwise they
seem to be the same.

So this patch requests the full_headers and deals with lowercasing the
desired headers.  This works for me as long as both
https://github.com/saltstack/salt/commit/8999148da0a96145f289d1049802e49c5d862339
and
https://github.com/saltstack/salt/commit/2a372054d7613779667186fc1d0dc7789cc5ad87
are applied.