Pass 'sock_dir' for all 'salt.utils.cloud.fire_event()' calls (#34806)
If no 'sock_dir' value is passed into the 'fire_event()' function it
would default to a value that ignores the 'base_dir' value set in the
master config file. We need to explicitly pass the 'sock_dir' value when
calling 'fire_event()' so that the correct directory is used.