Don't allow mercurial states to return True with errors
Fixes #36553

The hg.present state was not checking for errors for any of the calls
to the hg.py execution module functions. All of the hg.py execution
module functions were using cmd.run, instead of cmd.run_all, which
allowed for hg.py execution module functions to log an error at the CLI
but the hg.present state would return True, even though there were
problems executing the module functions.

This change adds try/except blocks around the calls to the mercurial
execution module functions and retuns False when a CommandExecutionError
is raised by the module. The module has been changes to use cmd.run_all
instead of cmd.run in order to check for the retcode of the underlying
mercurial calls and raises a CommandExecutionError is the retcode != 0.