Avoid errors when sudo_user is set
When sudo_user is used, the job is forked into a new pid when sudo is
executed. If a state then tries itself to execute states (such as when a
state invokes state.single), the fact that the forked PID was different
from the PID from which the job was started will cause Salt to refuse to
run the state.

This avoids this by conditionally setting concurrent to True when
invoking states from within states.