Pass in client_args when calling influxdb execution module funcs
Fixes #37996

There are several functions in the influxdb_user.py state module
that call functions in the influx.py execution module that
rely on passing **client_args to connect to the database.

This change adds the **client_args option to the influxbd_user state
to pass through these connection options to the execution module calls.