added docs about LXC container templates
In current docs it is not clear which templates are available and also
which options these have. Added references to the docs that help the
user find the list of templates as well as the options.

My local sphinx version is too old so I could not test the HTML output,
please verify syntax is correct.