Artifactory State: Only wrap main function call to module in try/except and wrap exc comment in str()
Some of the exceptions coming back from the execution module are formatted is different ways
coming from the URLError. When the state tries to return the error in the comment, we get a TypeError:

TypeError: Cannot serialize CommandExecutionError('Error: <urlopen error [Errno -2] Name or service not known>',)

Type casting the error message into a string helps prevent this from happening and gives the error
message back to the state.