states: sqlite3: fix table_present with multi-line schema
With a sqlite3 database containing a table with this schema:

  CREATE TABLE `user` (
  `id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL
  , `login` TEXT NOT NULL)

using `sqlite3.table_present` with this definition:

  user:
    sqlite3.table_present:
      - db: /tmp/sqlite.db
      - schema: |
          CREATE TABLE `user` (
          `id` INTEGER PRIMARY KEY AUTOINCREMENT NOT NULL
          , `login` TEXT NOT NULL)

returns an error indicating that the schemas do not match. That's
because the `schema` variable in the YAML ends with a newline, while the
schema returned by python-sqlite3 doesn't.

This commit strips leading and trailing whitespace from the YAML schema.