JunOS:  Add alive function to proxy
New in Nitrogen is a proxy_keep_alive option to poll the connection
periodicaly to keep it alive.  This is predicated on having an alive function
in the proxy code to see if the connection still exists.  This code
implements the alive function so that the proxy_keep_alive can work.