Fix creation of certain k8s secrets
Kubernetes secrets are internally represented as a dictionary
with strings as data type for both keys and values.

Ensure the objects we send to the API server respect this requirement
otherwise an internal error will take place on the server.

Also rely on kubernetes response to show the data changed, this fixes
issues when creating secrets from local files.

Signed-off-by: Flavio Castelli <fcastelli@suse.com>