Multiple documentation fixes
The master config 'engines' option takes a list of engines
but will accept a single engine instead of the list but adds
a warning that it requires a list. The documentation for some
of the engine modules pass a single engine instead of a list
of engines. This commit fixes the documentation of these engine
modules to give correct usage of the 'engines' option.

Fixes #42333

Additionally, A few of the documentation pages were missing
code-blocks where they were needed to ensure proper formatting.
This commit adds the code blocks where they were missing in the
engine modules documentation.