This block isn't necessary
If a location isn't passed to list_nodes_full, we can just use get_location().
Of no location is set in the provider, it will just use get_location().

The problem with the lookup if there are profiles, is the way that the
provider/driver was changed in the last release, if it just uses the profile
directly, it will fail when trying to do a full list.

Then for multiple locations being called, each provider has to have a location
in it, and will be called once, as it stands now, if you have multiple
providers, minions from each provider could be doubled up with this lookup.

This change should meet all the needs that this if statement was added for, and not fail in 2017.7