Handle VPC/Subnet ID not found errors in boto_vpc module
If a VPC or Subnet ID is not found when calling functions that are
supposed to be checking for vpc/subnet ID existence, the return
should be consistent by returning booleans/None instead of returning
the NotFound error from AWS.

The surrounding code blocks indicate that this is expected as well
as unit test assertions. The moto library had a bug in it where
it wasn't raising "x.NotFound" errors when it should have been. The
latest version of moto has fixed this bug, causing our tests to fail
since the boto_vpc module is not handling the "x.NotFound" errors
separately from the generic BotoServerErrors.

This fixes the test failures in the branch tests that were caused by
upgrading the moto version to the latest release.