Revert PR 41278
This PR should not have been merged, as there are very good reasons for
keeping the mongodb_database and mongodb_user state modules separate.
Additionally, role management was already supported at the time the PR
was merged.

Since this PR has not been in an official release of Salt, I'm reverting
it now to prevent any confusion.