virt module: don't force default connection URI
libvirt open() function handles None connection URI. In such a case it
tries all the hypervisors and connects to the first possible one.
Forcing 'qemu:///system' URI in the virt module was wrong since it
forced using the QEMU system session even on minions that don't have it.
With None as default, the user has nothing to do to connect to
hypervisors like Xen.

While at it, remove the use of __opts__ to access the 'virt.connect'
property, use config.get instead.