pillar_ldap: Fix cryptic errors when config file fails to load
This fixes an UnboundLocalError when the jinja templating or yaml
loading fails. It also re-adds some previously-deleted crucial
documentation regarding the `search_order` option.

Finally, it fixes some incorrectly-formatted RST and makes the result a
example python data structure.

We still need "split mode" documentation, but this was left undone with
a TODO telling the user to reference a docstring that isn't visible in
the docs.