remove uninitialized use of "current" variable
When modules.alternatives.check_installed() returns True, we know
that the default for this link group has been set to "path", which
we were passed in the first place. There's no need to fetch the
current state of the link group in that case (modulo race conditions
by concurrent changes to the link group, but there's no way to
catch that).
Fetching of the "current" state was already removed in the previous
commit, but one use of the "current" variable had been forgotten.