virt: fix disk targets before diffing
Ensure the disk target devices are unique within the disks in the new
list before diffing them. Libvirt will reorder the disks based on their
targets. If the users gives a disk order that is not consistent with the
currently running VM, then it's his fault giving us wrong data.