btrfs: add option to not set subvolumes as default
When we want to create a subvolume and set is as default, we can
use the set_default parameter. The current code will check, for
already created subvolumes, if it is actually marked as default,
and if not it will be set inside the state.

For composability with other states, we sometimes want to avoid
the re-set of the default flag for already present subvolumes, as
can be that this was change in a later action by another state.

(cherry picked from commit e95680800e5b3cee4a9a3f03cce9faf49f9a4497)