Improve the Capirca and related modules for ACL config generation
Since I added these modules in 2017.7.0, Capirca has been released on PyPI,
now available to be installed as a regular Python library.

Besides that, in the past there were various issues due to the way the modules
are interacting with the underlying library, solved by adding a hardcoded list
for the policy term fields. This has caused some issues in the past, and had
to be patches, e.g., 9bc777ba3458890ab36c24912cd7284ac0124ab1 or
1e74141cc08b704cc9d289f02eba2f009ce523d1.
While the changes referenced above are not particularly big, it's tedious to
always keep an eye on Capirca and be aware of what's newly added or changed.
Besides, this may have caused issues for users in the past when using other
Capirca versions.