Fix broken sdb.get_or_set_hash for Hashicorp Vault
When using Hashicorp Vault and SDB it's not possible to use `sdb.get_or_set_hash` because the `get` returns an `Exception` rather than None, which `get_or_set_hash` expects. Additionally, `sdb.get` returns a `CommandExecutionError` when it would be expected to return None.

Example response from `sdb.get_or_set_hash`:
```
Exception occurred in runner sdb.get_or_set_hash: Traceback (most recent call last):
  File "/usr/lib/python2.7/site-packages/salt/client/mixins.py", line 377, in low
    data['return'] = func(*args, **kwargs)
  File "/usr/lib/python2.7/site-packages/salt/runners/sdb.py", line 83, in get_or_set_hash
    return salt.utils.sdb.sdb_get_or_set_hash(uri, __opts__, length, chars, __utils__)
  File "/usr/lib/python2.7/site-packages/salt/utils/sdb.py", line 127, in sdb_get_or_set_hash
    ret = sdb_get(uri, opts, utils=utils)
  File "/usr/lib/python2.7/site-packages/salt/utils/sdb.py", line 46, in sdb_get
    return loaded_db[fun](query, profile=profile)
  File "/usr/lib/python2.7/site-packages/salt/sdb/vault.py", line 114, in get
    raise salt.exceptions.CommandExecutionError(e)
CommandExecutionError: 404 Client Error: Not Found
```

This PR fixes this by fixing `sdb.get` 

This PR does make the assumption that the correct behaviour for `get` when a value does not exist is to return None rather than a `CommandExecutionError`. I assume this because `CommandExecutionError` is an error-case - whereas the non-existence of a SDB URI is a perfectly valid state, and should be handled as a successful result, albeit with no found value. Currently, as it stands, because of this execution error it isn't possible to handle inside a Jinja template the non-existence of an SDB value in Vault.