tests: Add virt minion integration targets
The virt minion targets are used for integration testing of the virt
module. The targets are based on a container image with preinstalled
libvirt and qemu.

The salt source repository and configuration files are bind-mounted
within the container where a salt-minion instance is started.

The host's network stack (--network=host) is used to allow the
salt-minion inside the container to connect to the salt-master running
on the host.

Signed-off-by: Radostin Stoyanov <rstoyanov@fedoraproject.org>