modules/opkg: create restartcheck state if it's dir doesn't exist
This is a code cleanup change as the directory contains more than
the module.dep.{timestamp,md5sum} files, so the test was misleading.

The old file test was a remnant from when the only reboot sources were
kernel module changes (the files still get created unconditionally).

This is a change done by Ioan-Adrian Ratiu in NI salt repo: commit 505206

Signed-off-by: Cristian <cristian.hotea@ni.com>