Port #49697 and #49926 to master (#56893)
* Port #49697 to master

Fix set_static_all, always display requestmode and refactor
_change_state.

Display requestmode even if service is offline. This will keep parity
between older and newer nilrt distro.

Split the logic of _change_state function for older and newer disto.

Fix set_static_all. If there are no nameservers specified, we shouldn't
set any property for nameservers configuration to the service.

* Port #49926 to master.

Configure network interfaces without cable & add tests.

On NG we use connman as a network manager which consider a service, an
interface connected with a cable. The only way to set-up an interface
without a cable is through config files. This works also for services
because before we start the connamn daemon, we create a config file,
/var/lib/connman/interfaces.config with DHCP settings for each
interface.

* Require pyifaces, skip test if no pyifaces installed.

* Update unit test correspondingly to the module changes.

* Ran pre-commit hook

Co-authored-by: Alexandru Vasiu <alexandru.vasiu@ni.com>